[app.sirene-of-shame] ... saving of the last working stand.

+++ java/applications/serene-of-shame/hardware/sirene-of-shame_serial-port-controlled/Timer.cpp
@@ -1,9 +1,9 @@
 #include "Arduino.h"
 #include "Timer.h"
 
-bool Timer::isOver(int timeoutInMs) {
+bool Timer::isOver(long timeoutInMs) {
 
-  int currentTime = millis();
+  long currentTime = millis();
   return (currentTime - _timerStartTime) > timeoutInMs;
 }
